refactor: shell configurations idempotent
Also fixed some configurations errors pointed out by VirtualBox.

+++ Vagrantfile
@@ -7,16 +7,22 @@ INSTANCES = 2
 box_info = { name: 'generic/rocky8', version: '4.3.12' }
 
 PROVISION_PUPPET = <<~PUPPET
-  /usr/bin/dnf makecache
-  /usr/bin/dnf upgrade -y
-  /bin/rpm -Uvh https://yum.puppet.com/puppet7-release-el-8.noarch.rpm
-  /usr/bin/dnf -y install puppet-agent
-  /usr/bin/dnf autoremove -y
-  /usr/bin/dnf clean all
-  echo '*' > /etc/puppetlabs/puppet/autosign.conf
-  /opt/puppetlabs/bin/puppet resource host puppet.choria ensure=present ip=192.168.56.5 host_aliases=puppet
-  mkdir -p /etc/puppetlabs/facter/facts.d
-  echo "role=${1}" > /etc/puppetlabs/facter/facts.d/role.txt
+  ROLE_FILE=/etc/puppetlabs/facter/facts.d/role.txt
+  if [[ -f $ROLE_FILE ]]
+  then
+      echo 'Packages already installed, nothing to do'
+  else
+      /usr/bin/dnf makecache
+      /usr/bin/dnf upgrade -y
+      /bin/rpm -Uvh https://yum.puppet.com/puppet7-release-el-8.noarch.rpm
+      /usr/bin/dnf -y install puppet-agent
+      /usr/bin/dnf autoremove -y
+      /usr/bin/dnf clean all
+      echo '*' > /etc/puppetlabs/puppet/autosign.conf
+      /opt/puppetlabs/bin/puppet resource host puppet.choria ensure=present ip=192.168.56.5 host_aliases=puppet
+      mkdir -p /etc/puppetlabs/facter/facts.d
+      echo "role=${1}" > /etc/puppetlabs/facter/facts.d/role.txt
+  fi
 PUPPET
 
 Vagrant.configure('2') do |config|
@@ -27,8 +33,11 @@ Vagrant.configure('2') do |config|
     vmconfig.vm.box_version = box_info[:version]
     vmconfig.vm.hostname = 'puppet.choria'
     vmconfig.vm.network :private_network, ip: '192.168.56.5'
+    vmconfig.vm.network :forwarded_port, guest: 9100, host: 9100, id: 'Prometheus'
     vmconfig.vm.provider :virtualbox do |vb|
       vb.customize ['modifyvm', :id, '--memory', 3072]
+      vb.customize ['modifyvm', :id, '--graphicscontroller', 'vmsvga']
+      vb.customize ['modifyvm', :id, '--vram', '16']
     end
 
     vmconfig.vbguest.installer_options = { allow_kernel_upgrade: true }
@@ -40,19 +49,24 @@ Vagrant.configure('2') do |config|
     end
 
     vmconfig.vm.provision 'shell', inline: <<-SHELL
-      echo ">>> Installing PuppetServer..."
-      /bin/rpm -Uvh https://yum.puppet.com/puppetserver-release-el-8.noarch.rpm
-      dnf install -y puppetserver
+      if systemctl status puppetserver --no-pager &> /dev/null
+      then
+          echo 'Puppet already installed, nothing to do'
+      else
+          echo ">>> Installing PuppetServer..."
+          /bin/rpm -Uvh https://yum.puppet.com/puppetserver-release-el-8.noarch.rpm
+          dnf install -y puppetserver
 
-      echo ">>> Syncing environments to /etc/puppetlabs/code/environments..."
-      rsync -a --delete /vagrant/environments/ /etc/puppetlabs/code/environments/
+          echo ">>> Syncing environments to /etc/puppetlabs/code/environments..."
+          rsync -a --delete /vagrant/environments/ /etc/puppetlabs/code/environments/
 
-      echo ">>> Starting PuppetServer..."
-      systemctl start puppetserver
-      systemctl enable puppetserver
+          echo ">>> Starting PuppetServer..."
+          systemctl start puppetserver
+          systemctl enable puppetserver
 
-      sleep 10
-      systemctl status puppetserver --no-pager | head -5
+          sleep 10
+          systemctl status puppetserver --no-pager | head -5
+      fi
     SHELL
 
     vmconfig.vm.provision 'shell', inline: <<-SHELL
@@ -71,6 +85,8 @@ Vagrant.configure('2') do |config|
       vmconfig.vm.network :private_network, ip: format('192.168.56.%d', 9 + i)
       vmconfig.vm.provider :virtualbox do |vb|
         vb.customize ['modifyvm', :id, '--memory', 1024]
+        vb.customize ['modifyvm', :id, '--graphicscontroller', 'vmsvga']
+        vb.customize ['modifyvm', :id, '--vram', '16']
       end
 
       vmconfig.vbguest.installer_options = { allow_kernel_upgrade: true }
@@ -90,6 +106,7 @@ Vagrant.configure('2') do |config|
           --hiera_config=/etc/puppetlabs/code/environments/production/hiera.yaml \
           --modulepath=/etc/puppetlabs/code/environments/production/modules:/etc/puppetlabs/code/environments/production/site
       SHELL
+
       vmconfig.vm.provision 'shell', inline: <<-SHELL
         echo ">>> Running puppet agent to enroll with PuppetServer..."
         /opt/puppetlabs/bin/puppet agent -tv --waitforcert 30 --server puppet.choria
